+++ clumio/plugin_framework/clumio_policy_rule/const.go
@@ -5,11 +5,10 @@ package clumio_policy_rule
 const (
 	// Constants used by the resource model for the clumio_policy_rule Terraform resource. These
 	// values should match the schema tfsdk tags on the resource model struct in schema.go.
-	schemaName                 = "name"
-	schemaId                   = "id"
-	schemaCondition            = "condition"
-	schemaBeforeRuleId         = "before_rule_id"
-	schemaPolicyId             = "policy_id"
-	schemaOrganizationalUnitId = "organizational_unit_id"
-	schemaPolicyRules          = "policy_rules"
+	schemaName         = "name"
+	schemaId           = "id"
+	schemaCondition    = "condition"
+	schemaBeforeRuleId = "before_rule_id"
+	schemaPolicyId     = "policy_id"
+	schemaPolicyRules  = "policy_rules"
 )