+++ clumio/plugin_framework/clumio_user/resource_user_test.go
@@ -105,63 +105,6 @@ func TestAccResourceClumioUser(t *testing.T) {
 	})
 }
 
-// Test of the clumio_user resource with assigned_role and organizational_unit_ids schema attributes
-// populated. It tests the following scenarios:
-//   - Creates a user and verifies that the plan was applied properly.
-//   - Updates the user and verifies that the resource will be updated.
-func TestAccResourceClumioUserOURole(t *testing.T) {
-
-	// Run the acceptance test.
-	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { clumioPf.UtilTestAccPreCheckClumio(t) },
-		ProtoV6ProviderFactories: clumioPf.TestAccProtoV6ProviderFactories,
-		Steps: []resource.TestStep{
-			{
-				Config: getTestAccResourceClumioUserRoleIdAndOus(false),
-				ConfigPlanChecks: resource.ConfigPlanChecks{
-					PreApply: []plancheck.PlanCheck{
-						plancheck.ExpectNonEmptyPlan(),
-					},
-					PostApplyPreRefresh: []plancheck.PlanCheck{
-						plancheck.ExpectEmptyPlan(),
-					},
-					PostApplyPostRefresh: []plancheck.PlanCheck{
-						plancheck.ExpectEmptyPlan(),
-						plancheck.ExpectResourceAction(
-							"clumio_user.test_user", plancheck.ResourceActionNoop),
-					},
-				},
-				Check: resource.ComposeTestCheckFunc(
-					resource.TestMatchResourceAttr(
-						"clumio_user.test_user", "assigned_role",
-						regexp.MustCompile(assignedRoleBefore)),
-				),
-			},
-			{
-				Config: getTestAccResourceClumioUserRoleIdAndOus(true),
-				ConfigPlanChecks: resource.ConfigPlanChecks{
-					PreApply: []plancheck.PlanCheck{
-						plancheck.ExpectNonEmptyPlan(),
-					},
-					PostApplyPreRefresh: []plancheck.PlanCheck{
-						plancheck.ExpectEmptyPlan(),
-					},
-					PostApplyPostRefresh: []plancheck.PlanCheck{
-						plancheck.ExpectEmptyPlan(),
-						plancheck.ExpectResourceAction(
-							"clumio_user.test_user", plancheck.ResourceActionNoop),
-					},
-				},
-				Check: resource.ComposeTestCheckFunc(
-					resource.TestMatchResourceAttr(
-						"clumio_user.test_user", "assigned_role",
-						regexp.MustCompile(assignedRoleAfter)),
-				),
-			},
-		},
-	})
-}
-
 // Tests that an external deletion of a clumio_user resource leads to the resource needing
 // to be re-created during the next plan. NOTE the Check function below as it is utilized to delete
 // the resource using the Clumio API after the plan is applied.
@@ -335,19 +278,6 @@ func getTestAccResourceClumioUser(baseUrl string, name string, email string, upd
 	return fmt.Sprintf(testAccResourceClumioUser, baseUrl, name, email, assignedRole, orgUnitId)
 }
 
-// getTestAccResourceClumioUser returns the Terraform configuration for creating a clumio_user
-// resource using role_id and organizational_unit_ids instead of access_control_configuration.
-func getTestAccResourceClumioUserRoleIdAndOus(update bool) string {
-
-	orgUnitId := "clumio_organizational_unit.test_ou1.id"
-	assignedRole := assignedRoleBefore
-	if update {
-		orgUnitId = "clumio_organizational_unit.test_ou2.id"
-		assignedRole = assignedRoleAfter
-	}
-	return fmt.Sprintf(testAccResourceClumioUserRoleIdAndOus, assignedRole, orgUnitId)
-}
-
 // testAccResourceClumioUser is the Terraform configuration for a basic clumio_user resource.
 const testAccResourceClumioUser = `
 provider clumio{
@@ -385,29 +315,5 @@ provider clumio{
 resource "clumio_user" "test_user" {
   full_name = "acceptance-test-user"
   email = "test@clumio.com"
-}
-`
-
-// testAccResourceClumioUserRoleIdAndOus is the Terraform configuration for creating a clumio_user
-// resource using role_id and organizational_unit_ids instead of access_control_configuration.
-const testAccResourceClumioUserRoleIdAndOus = `
-provider clumio{
-}
-
-resource "clumio_organizational_unit" "test_ou1" {
-  name = "test_ou1"
-  description = "test-ou-1"
-}
-
-resource "clumio_organizational_unit" "test_ou2" {
-  name = "test_ou2"
-  description = "test-ou-2"
-}
-
-resource "clumio_user" "test_user" {
-  full_name = "acceptance-test-user"
-  email = "test@clumio.com"
-  assigned_role = "%s"
-  organizational_unit_ids = [%s]
 }
 `