+++ examples/guides/cross_region.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     clumio = {
       source  = "clumio-code/clumio"
-      version = "~>0.10.0"
+      version = "~>0.11.0"
     }
     aws = {}
   }
@@ -17,11 +17,21 @@ provider "clumio" {
 # Instantiate two AWS providers with different regions
 provider "aws" {
   region = "us-west-2"
+  default_tags {
+    tags = {
+      "Vendor" = "Clumio"
+    }
+  }
 }
 
 provider "aws" {
   alias  = "east"
   region = "us-east-1"
+  default_tags {
+    tags = {
+      "Vendor" = "Clumio"
+    }
+  }
 }
 
 # Retrieve the effective AWS account ID