+++ examples/resources/clumio_aws_connection/resource.tf
@@ -2,5 +2,4 @@ resource "clumio_aws_connection" "example" {
   account_native_id      = "aws_account_id"
   aws_region             = "aws_region"
   description            = "description"
-  organizational_unit_id = "organizational_unit_id"
 }